実行環境 sandbox を整備する(2) refs #3

+++ sandbox/run_docker.sh
@@ -1,21 +1,24 @@
 #!/bin/bash
-#- 自分の sandbox コンテナを立ち上げる (共有ディレクトリ: ~/temp/)
+#- 自分の sandbox コンテナを立ち上げる (共有ディレクトリ: ~/sandbox/)
 
 set -e
 
 user=`id -nu`
-home="/home/${user}"
-name="${user}"
+#name="${user}"
+name="sandbox"
+home=${HOME}  #- ホストとDockerコンテナで同じhome directoryとする
+dir_shared=${HOME}/sandbox
+mkdir -p ${dir_shared}
 
 #args="-e http_proxy=${HTTP_PROXY} -e HTTPS_PROXY=${HTTP_PROXY}"
 # HTTPS_PROXY is used by pip
 args=""
 
-docker run ${args} -v ${home}/temp:${home}/temp \
+docker run ${args} -v ${dir_shared}:${dir_shared} \
        --rm -itd --name ${name} sandbox:${user}
 
 docker cp myconf/.bashrc ${name}:${home}/
-docker cp ${HOME}/.gitconfig ${name}:${home}/
+[ -f ${HOME}/.gitconfig ] && docker cp ${HOME}/.gitconfig ${name}:${home}/
 
 docker exec -it ${name} /bin/bash
